Function captain::init

pub fn init(
    kernel_mmi_ref: MmiRef,
    identity_mapped_pages: NoDrop<EarlyIdentityMappedPages>,
    bsp_initial_stack: NoDrop<Stack>,
    ap_start_realmode_begin: VirtualAddress,
    ap_start_realmode_end: VirtualAddress,
    ap_gdt: VirtualAddress,
    rsdp_address: Option<PhysicalAddress>
) -> Result<(), &'static str>

Initialize the Captain, which is the main crate that “steers the ship” of Theseus.

This does the rest of the initialization procedures so that the OS can continue running and do actual useful work.

Arguments

kernel_mmi_ref: a reference to the kernel’s memory management info.

identity_mapped_pages: the memory containing the identity-mapped content, which must not be dropped until all APs are finished booting.

bsp_initial_stack: the stack currently in use for running this code, which must not be dropped for the entire execution of the initial bootstrap task.

ap_start_realmode_begin: the start bound (inclusive) of the AP’s realmode boot code.

ap_start_realmode_end: the end bound (exlusive) of the AP’s realmode boot code.

ap_gdt: the virtual address of the GDT created for the AP’s realmode boot code.

rsdp_address: the physical address of the RSDP (an ACPI table pointer), if available and provided by the bootloader.